ndk-examples: Bump `android-activity`/`android-logger` to resolve crashes
Both `log` and `android-activity` seem to be creating vectors or slices
from NULL pointers, which was already invalid but the standard library
now asserts and panics on.  Since both have been fixed, upgrade all
dependencies to get the local and CI runs working again.

To keep duplicate dependencies to a minimum, upgrade `ndk` to `0.9`
as well which now uses `OwnedFd`/`BorrowedFd` to accurately represent
ownership transitions.  This allows us to convert the `looper` example
to safe `rustix` bindings.